requirements.txt
django==1.7
mysql-python
django-suit
django-debug-toolbar==1.3.2
pytils==0.3
pillow
xlwt==1.0.0
django-annoying==0.8.3
django-imagekit==3.2.7
django-widget-tweaks==1.4.1
django-inplaceedit==1.4.1
django-inplaceedit-bootstrap==0.2.1
django-ckeditor==4.5.1
django-cleanup==0.4.2
